[Android/Feat] 스위치 상태 정보 조회 API 연동 S12P31D103-206

+++ Android/app/src/main/java/com/hogumiwarts/lumos/ui/screens/routine/routineDeviceList/devicecontrolscreen/PreviewAirPurifierScreenContent.kt
@@ -52,7 +52,7 @@ fun PreviewAirPurifierScreenContent(
     val powerState by viewModel.powerState.collectAsState()
 
     LaunchedEffect(deviceId) {
-        viewModel.sendIntent(AirpurifierIntent.LoadAirpurifierStatus(deviceId))
+        viewModel.sendIntent(AirpurifierIntent.LoadAirpurifierStatus(deviceId.toLong()))
     }
 
 
@@ -123,7 +123,7 @@ fun PreviewAirPurifierScreenContent(
                 Switch(
                     checked = checked,
                     onCheckedChange = {
-                        viewModel.sendIntent(AirpurifierIntent.ChangeAirpurifierPower(deviceId, it))
+                        viewModel.sendIntent(AirpurifierIntent.ChangeAirpurifierPower(deviceId.toLong(), it))
                         checked = it
                     },
                     colors = SwitchDefaults.colors(
